incremental
OK, now the convolution against simulation derivative is working now.
Still gives about 50% of shots are good at IPM index 30, but still, this is conservative since the leading
edge is actually easier to see than the falling and so using the diff between the two to tell which are good shots
is maybe a bit more restrictive than need be.